requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-docx==1.0.1
openpyxl==3.1.2
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
PyMuPDF==1.23.8
SpeechRecognition==3.10.0
ffmpeg-python==0.2.0
pytesseract==0.3.10
google-genai>=0.1.0
python-dotenv==1.0.0